+++ ChangeLog.md
@@ -5,13 +5,27 @@ This project's release branch is `master`. This log is written from the perspect
 ## Unreleased
 
 * Bump
-  * all-cabal-hashes ("Update from Hackage at 2020-11-05T15:13:14Z")
+  * all-cabal-hashes ("Update from Hackage at 2020-11-06T16:55:57Z")
   * which to 0.2.0.0
   * gargoyle to 0.1.1.0
   * gargoyle-postgresql to 0.2.0.1
   * gargoyle-postgresql-nix to 0.3.0.0
   * aeson-gadt-th to 0.2.4
-  * reflex-ghci to 0.1.5.0
+  * reflex-ghci to 0.1.5.1
+  * constraints-extras to 0.3.0.2
+  * some to 1.0.1
+  * prim-uniq to 0.2
+  * aeson-gadt-th to 0.2.4
+  * dependent-map to 0.4
+  * dependent-sum to 0.7.1.0
+  * dependent-sum-template to 0.1.0.3
+  * dependent-sum-aeson-orphans to 0.3.0.0
+  * reflex to 0.7.2.0
+  * reflex-vty to 0.1.4.1
+  * patch to 0.0.3.2
+  * reflex-dom to 0.6.1.0
+  * reflex-fsnotify to 0.2.1.2
+  * reflex-process to 0.3.1.0
 * New Pins
   * gargoyle-postgresql-connect 0.1.0.0